MFB-1676: give CESN its own white label directory
CESN's 20 remaining programs move out of programs/programs/co/energy_calculator/
into white_labels/cesn/<program>/. CESN is a non-geographic white label with its
own Program rows; nesting it inside CO's tree said otherwise.

cesn_leap and cesn_cowap are not here: they moved into cross_white_label/liheap/
and .../weatherization/ in part 1, because each is one implementation of a
federal benefit. cesn_cpcr does come here, since CO's property credit rebate is
a state program rather than a family.
