+++ programs/programs/white_labels/cesn/bheap/calculator.py
@@ -1,14 +1,14 @@
 from programs.framework.base import Eligibility, ProgramCalculator
-from programs.programs.cross_white_label.liheap.cesn import EnergyCalculatorEnergyAssistance
-from programs.programs.co.energy_calculator.energy_outreach.calculator import EnergyCalculatorEnergyOutreach
 from programs.programs.co.energy_calculator.util import has_renter_expenses
+from programs.programs.cross_white_label.liheap.cesn import EnergyCalculatorEnergyAssistance
+from programs.programs.white_labels.cesn.eoc.calculator import EnergyCalculatorEnergyOutreach
 from programs.programs.cross_white_label.weatherization.cesn import (
     EnergyCalculatorWeatherizationAssistance,
 )
 
 
-class EnergyCalculatorElectricityAffordabilityXcel(ProgramCalculator):
-    program_code = "cesn_xceleap"
+class EnergyCalculatorElectricityAffordabilityBlackHills(ProgramCalculator):
+    program_code = "cesn_bheap"
     amount = 1
     dependencies = [
         *EnergyCalculatorEnergyAssistance.dependencies,
@@ -20,8 +20,9 @@ class EnergyCalculatorElectricityAffordabilityXcel(ProgramCalculator):
         "cesn_leap",
         "cesn_eoc",
         "cesn_cowap",
+        "cesn_care",
     ]
-    electricity_providers = ["co-xcel-energy"]
+    electricity_providers = ["co-black-hills-energy"]
 
     def household_eligible(self, e: Eligibility):
         # eligible for another program
@@ -32,7 +33,7 @@ class EnergyCalculatorElectricityAffordabilityXcel(ProgramCalculator):
                 has_another_program = True
         e.condition(has_another_program)
 
-        # has gas provider
+        # has electric provider
         e.condition(self.screen.energy_calculator.has_electricity_provider(self.electricity_providers))
 
         # no renters without expenses
