+++ programs/programs/white_labels/cesn/cngba/calculator.py
@@ -1,14 +1,14 @@
 from programs.framework.base import Eligibility, ProgramCalculator
 from programs.programs.co.energy_calculator.util import has_renter_expenses
 from programs.programs.cross_white_label.liheap.cesn import EnergyCalculatorEnergyAssistance
-from programs.programs.co.energy_calculator.energy_outreach.calculator import EnergyCalculatorEnergyOutreach
+from programs.programs.white_labels.cesn.eoc.calculator import EnergyCalculatorEnergyOutreach
 from programs.programs.cross_white_label.weatherization.cesn import (
     EnergyCalculatorWeatherizationAssistance,
 )
 
 
-class EnergyCalculatorElectricityAffordabilityBlackHills(ProgramCalculator):
-    program_code = "cesn_bheap"
+class EnergyCalculatorNaturalGasBillAssistance(ProgramCalculator):
+    program_code = "cesn_cngba"
     amount = 1
     dependencies = [
         *EnergyCalculatorEnergyAssistance.dependencies,
@@ -20,9 +20,8 @@ class EnergyCalculatorElectricityAffordabilityBlackHills(ProgramCalculator):
         "cesn_leap",
         "cesn_eoc",
         "cesn_cowap",
-        "cesn_care",
     ]
-    electricity_providers = ["co-black-hills-energy"]
+    gas_providers = ["co-colorado-natural-gas"]
 
     def household_eligible(self, e: Eligibility):
         # eligible for another program
@@ -33,8 +32,8 @@ class EnergyCalculatorElectricityAffordabilityBlackHills(ProgramCalculator):
                 has_another_program = True
         e.condition(has_another_program)
 
-        # has electric provider
-        e.condition(self.screen.energy_calculator.has_electricity_provider(self.electricity_providers))
+        # has gas provider
+        e.condition(self.screen.energy_calculator.has_gas_provider(self.gas_providers))
 
         # no renters without expenses
         e.condition(has_renter_expenses(self.screen))
