+++ programs/programs/white_labels/cesn/poipp/calculator.py
@@ -1,7 +1,7 @@
 from programs.framework.base import Eligibility, ProgramCalculator
 from programs.programs.co.energy_calculator.util import has_renter_expenses
 from programs.programs.cross_white_label.liheap.cesn import EnergyCalculatorEnergyAssistance
-from programs.programs.co.energy_calculator.energy_outreach.calculator import EnergyCalculatorEnergyOutreach
+from programs.programs.white_labels.cesn.eoc.calculator import EnergyCalculatorEnergyOutreach
 from programs.programs.cross_white_label.weatherization.cesn import (
     EnergyCalculatorWeatherizationAssistance,
 )
