+++ programs/programs/white_labels/cesn/xceleap/calculator.py
@@ -1,14 +1,14 @@
 from programs.framework.base import Eligibility, ProgramCalculator
-from programs.programs.co.energy_calculator.util import has_renter_expenses
 from programs.programs.cross_white_label.liheap.cesn import EnergyCalculatorEnergyAssistance
-from programs.programs.co.energy_calculator.energy_outreach.calculator import EnergyCalculatorEnergyOutreach
+from programs.programs.white_labels.cesn.eoc.calculator import EnergyCalculatorEnergyOutreach
+from programs.programs.co.energy_calculator.util import has_renter_expenses
 from programs.programs.cross_white_label.weatherization.cesn import (
     EnergyCalculatorWeatherizationAssistance,
 )
 
 
-class EnergyCalculatorNaturalGasBillAssistance(ProgramCalculator):
-    program_code = "cesn_cngba"
+class EnergyCalculatorElectricityAffordabilityXcel(ProgramCalculator):
+    program_code = "cesn_xceleap"
     amount = 1
     dependencies = [
         *EnergyCalculatorEnergyAssistance.dependencies,
@@ -21,7 +21,7 @@ class EnergyCalculatorNaturalGasBillAssistance(ProgramCalculator):
         "cesn_eoc",
         "cesn_cowap",
     ]
-    gas_providers = ["co-colorado-natural-gas"]
+    electricity_providers = ["co-xcel-energy"]
 
     def household_eligible(self, e: Eligibility):
         # eligible for another program
@@ -33,7 +33,7 @@ class EnergyCalculatorNaturalGasBillAssistance(ProgramCalculator):
         e.condition(has_another_program)
 
         # has gas provider
-        e.condition(self.screen.energy_calculator.has_gas_provider(self.gas_providers))
+        e.condition(self.screen.energy_calculator.has_electricity_provider(self.electricity_providers))
 
         # no renters without expenses
         e.condition(has_renter_expenses(self.screen))
